+++ strict.d.ts
@@ -16,4 +16,56 @@ declare global {
       selector: S,
     ): [E] extends [never] ? never : E | null
   }
+
+  interface HTMLElement {
+    querySelector<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : E | null
+
+    querySelectorAll<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : NodeListOf<E>
+
+    closest<S extends string, E extends StrictlyParseSelector<S, HTMLElement>>(
+      selector: S,
+    ): [E] extends [never] ? never : E | null
+  }
+
+  interface Document {
+    querySelector<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : E | null
+
+    querySelectorAll<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : NodeListOf<E>
+  }
+
+  interface DocumentFragment {
+    querySelector<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : E | null
+
+    querySelectorAll<
+      S extends string,
+      E extends StrictlyParseSelector<S, HTMLElement>,
+    >(
+      selector: S,
+    ): [E] extends [never] ? never : NodeListOf<E>
+  }
 }